登录界面UI 修改

+++ Assets/MirrorMarketSDK/Runtime/Models/Request/Marketplace/CancelNftListOnMarketplaceRequest.cs
@@ -6,7 +6,7 @@ namespace MirrorworldSDK.Models
     {
         [JsonProperty("mint_address")] public string MintAddress;
 
-        [JsonProperty("price")] public decimal Price;
+        [JsonProperty("price")] public float Price;
 
         [JsonProperty("confirmation")] public string Confirmation;
     }
