+++ Assets/MirrorMarketSDK/Runtime/Models/Request/Marketplace/ListNftOnMarketplaceRequest.cs
@@ -6,6 +6,6 @@ namespace MirrorworldSDK.Models
     {
         [JsonProperty("mint_address")] public string MintAddress;
 
-        [JsonProperty("price")] public decimal Price;
+        [JsonProperty("price")] public float Price;
     }
 }
